refactor: css append
