+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## 1.0.0
 
 - chore: update dependencies
-- test: use nyc
+- test: use nyc test coverage
+- refactor: css append
 
 ## 0.1.5
