+++ lib/200.js
@@ -2,24 +2,24 @@ const { readFileSync } = require('fs')
 const { join } = require('path')
 const mime = require('mime')
 
-function inject(filePath) {
-  if (mime.getType(filePath) !== 'text/html') {
-    return ''
-  }
-  return `<script>${readFileSync(join(__dirname, 'inject.js'))}</script>`
-}
-
 module.exports = (res, filePath) => {
   const file = readFileSync(filePath)
+  const type = mime.getType(filePath)
+  const inject = readFileSync(join(__dirname, 'inject.js'))
 
   res.writeHead(200, {
-    'Content-Type': mime.getType(filePath),
+    'Content-Type': type,
     'Cache-Control': 'private, no-cache, no-store, must-revalidate',
     Expires: '-1',
     Pragma: 'no-cache',
     'Access-Control-Allow-Origin': '*',
   })
+
   res.write(file)
-  res.write(inject(filePath))
+
+  if (type === 'text/html') {
+    res.write(`<script>${inject}</script>`)
+  }
+
   res.end()
 }
