+++ lib/inject.js
@@ -24,10 +24,14 @@
           var rel = elem.rel
           var href = elem.href
           if (rel === 'stylesheet') {
-            head.removeChild(elem)
+            var link = document.createElement('link')
+            link.rel  = 'stylesheet'
             href = href.split('?')[0]
-            elem.href = href + '?_=' + (+new Date())
-            head.appendChild(elem)
+            link.href = href + '?_=' + (+new Date())
+            head.appendChild(link)
+            setTimeout(function () {
+              head.removeChild(elem)
+            })
           }
         }
       }
