+++ test/console.js
@@ -1,0 +1,3 @@
+/* global window */
+
+window.console.log('Pavane')
