+++ test/dir/index.html
@@ -1,4 +1,4 @@
-<script src="../index.js"></script>
-<link rel="stylesheet" href="../index.css" />
+<script src="../console.js"></script>
+<link rel="stylesheet" href="../style.css" />
 <h1>Pavane</h1>
 <img src="../41502630.jpeg" />
