--- test/index.css
@@ -1,1 +1,0 @@
-body { background: #eee }
